Create tag in the question model.
This allows questions to modify their designated tag instead of making a new one every time it's saved. The question still, however, sends the request over the wire via the AJAX call. Needs refactor to check if the tag is at all different from the last time it sent the data.