Nullify sequence number on welcome page.
This hides the pagination when manually revisiting the welcome page. It is a more proper way of handling state because all of the chapter's routes must update the pagination nav via the pagination nav service. As of this commit, that is welcome, question, and results routes.

May need refactor by moving updating pagination to something that automatically does it upon any transition within the index.chapter route. The member, the chapter, and the sequence number are all available before going into and moving between children of index.chapter route.