Refactor: In pagination nav service, filter progresses for only the progress for this chapter, and update only that one's sequence number, and update pagination according to that one's sequence number thereafter.
This is still checking if there is a chapter_progress, but that should be guaranteed at this point because the chapter sets it if it does not exist. Needs refactor, but it is left in for now as to illustrate the similarity to the chapter route's evaluating and/or updating of the sequence number.