Refactor: Index route handles adding local storage progress markers. Chapter route handles redirecting according to progress marker.
Index route follows this precedence to seed progress markers:

1) use API member data's progress marker
2) use local storage's progress marker
3) create new progress marker

Chapter route is guaranteed a progress marker on the member for this chapter. It only redirects according to progress marker.

Pagination-nav service is guaranteed a progress marker on the member for this chapter. It updates the member's progress marker with the passed in sequence_num, and updates local storage with member's progress marker.