Get all of the member's tags from the server once, and then filter those according to the route's needs.
Routes were requesting tags from the server and then acting on that data, but it is possible the promise was not resolved before actions were taken on that data. This ensures that the promise is resolved at the index route level, and reduces the fetching of the tags to one call. Since the only way to CRUD tags is in routes nested under the index route, they won't get out of sync with the server.