Don't send the tag over the network on saveTag().
This allows for the server and front-end to become out of sync since a selection can be stored in local storage and appear as if it has been sent, but it has not. A way to ensure that the front-end and server are in sync is to have all tags only found in local storage to be automatically sent over the network.