Mock API handles 'patch' requests on members. Needs refactor.
This is needed because the app acts as if it is updating the member to the server.

It is only returning an ID to suppress the error "An adapter cannot assign a new id to a record that already has an id." This issue needs to be addressed, but it does not currently affect the app.