Fix: Send empty string "" for response value to Solarium for deleted responses. Needs refactor.
This will eventually be reverted back to an empty array once Solarium can handle arrays for the response value.

If the server starts returning empty strings (or eventually an empty array) for a tag's answer, the app won't know that the app "deleted" it before. In other words, a skipped tag, a tag that was never answered, and a tag that returned from the server that was previously deleted all look exactly the same. However, the server SHOULD NOT return such a value. A tag's answer is always an array in summer-app.